+++ pkg/lang/codegen.go
@@ -1395,6 +1395,21 @@ func (g *irGen) call(b *strings.Builder, c *Call) (string, error) {
 					}
 				}
 				return fmt.Sprintf("%d", res), nil
+			case "isspace":
+				res := 0
+				if v != "" {
+					all := true
+					for _, r := range v {
+						if !unicode.IsSpace(r) {
+							all = false
+							break
+						}
+					}
+					if all {
+						res = 1
+					}
+				}
+				return fmt.Sprintf("%d", res), nil
 			case "islower":
 				res := 0
 				hasCased := false